+++ daydreaming/src/main/java/com/brainydroid/daydreaming/background/DailySequenceService.java
@@ -21,6 +21,7 @@ import com.brainydroid.daydreaming.network.SntpClient;
 import com.brainydroid.daydreaming.network.SntpClientCallback;
 import com.brainydroid.daydreaming.sequence.Sequence;
 import com.brainydroid.daydreaming.sequence.SequenceBuilder;
+import com.brainydroid.daydreaming.ui.dashboard.SettingsActivity;
 import com.brainydroid.daydreaming.ui.sequences.PageActivity;
 import com.google.inject.Inject;
 
@@ -186,6 +187,12 @@ public class DailySequenceService extends RoboService {
     private synchronized void expireProbe(int probeId) {
         Logger.v(TAG, "Expiring probe");
         Sequence probe = sequencesStorage.get(probeId);
+        if (probe == null) {
+            Logger.v(TAG, "Probe not in DB any more, probably answered+synced+flushed. " +
+                    "No need to expire it.");
+            return;
+        }
+
         String status = probe.getStatus();
         if (status.equals(Sequence.STATUS_PENDING)) {
             probe.setStatus(Sequence.STATUS_RECENTLY_MISSED);
@@ -245,25 +252,19 @@ public class DailySequenceService extends RoboService {
 
         int flags = 0;
 
-
-        // Should we flash the LED?
-        if (sharedPreferences.getBoolean("notification_blink_key", true)) {
-            Logger.v(TAG, "Activating lights");
-            flags |= Notification.DEFAULT_LIGHTS;
-        }
-
-        // If this is NOT an MQ, we can vibrate and beep.
+        // If this is NOT an MQ, we can flash lights and vibrate.
+        // Beeping is dealt with further down (custom sound on the notification object).
         if (!sequenceType.equals(Sequence.TYPE_MORNING_QUESTIONNAIRE)) {
-            // Should we vibrate?
-            if (sharedPreferences.getBoolean("notification_vibrator_key", true)) {
-                Logger.v(TAG, "Activating vibration");
-                flags |= Notification.DEFAULT_VIBRATE;
+            // Should we flash the LED?
+            if (sharedPreferences.getBoolean(SettingsActivity.NOTIF_BLINK, true)) {
+                Logger.v(TAG, "Activating lights");
+                flags |= Notification.DEFAULT_LIGHTS;
             }
 
-            // Should we beep?
-            if (sharedPreferences.getBoolean("notification_sound_key", true)) {
-                Logger.v(TAG, "Activating sound");
-                flags |= Notification.DEFAULT_SOUND;
+            // Should we vibrate?
+            if (sharedPreferences.getBoolean(SettingsActivity.NOTIF_VIBRATION, true)) {
+                Logger.v(TAG, "Activating vibration");
+                flags |= Notification.DEFAULT_VIBRATE;
             }
         }
 
@@ -293,10 +294,12 @@ public class DailySequenceService extends RoboService {
 
         Notification notification = notificationBuilder.build();
 
-        // How to beep?
-        if (sharedPreferences.getBoolean("notification_sound_key", true)) {
-            Logger.v(TAG, "Adding custom sound");
-            notification.sound = Uri.parse("android.resource://" + "com.brainydroid.daydreaming" + "/" + R.raw.notification);
+        // Should we beep?
+        if (!sequenceType.equals(Sequence.TYPE_MORNING_QUESTIONNAIRE)
+                && sharedPreferences.getBoolean(SettingsActivity.NOTIF_SOUND, true)) {
+            Logger.v(TAG, "Activating beep for notification, custom sound");
+            notification.sound = Uri.parse("android.resource://" + "com.brainydroid.daydreaming" +
+                    "/" + R.raw.notification);
         }
 
         // Get a proper id for the notification, to replace the right notifications